test: fix RecruitmentAdminPageRenderer stub leak via fixture (#563)
The renderer test defined collaborator stub classes (RecruitmentAdminPage,
RecruitmentSettings, the readers, list-tables, Capabilities, AdminUI) in
file-scope namespace blocks. PHPUnit includes every *Test.php during suite
discovery in the parent process, so those guarded stubs were declared there
BEFORE the real classes autoloaded — poisoning the real recruitment classes
for ~70 other tests (@runTestsInSeparateProcesses only isolates method
execution, not file inclusion). Moved the stubs to a non-discovered fixture
under tests/fixtures/ and require_once it from setUp (runs only in the
isolated child). No coverage lost; RecruitmentAdminPageRenderer stays ~98%.
